+++ Js2IL.Tests/ControlFlow/JavaScript/ControlFlow_ForLoop_LetClosureCapture.js
@@ -1,0 +1,9 @@
+const fns = [];
+
+for (let i = 0; i < 3; i++) {
+  fns.push(() => i);
+}
+
+console.log(fns[0]());
+console.log(fns[1]());
+console.log(fns[2]());